Refactor Tag management into new Tags class
The Tag class has been decommissioned and replaced with the new Tags class, which improves the tag handling system. Key static methods for tag validation, retrieval, and formatting have been ported over and modified to support new features. Notably, functionality to create a hierarchical tag tree has been introduced, offering a structured representation of tag relationships. This change centralizes tag-related operations, streamlines the API, and prepares the codebase for more advanced tag manipulation in the future.
